# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  },
                  "required": [
                    "access_token"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Authenticate a user and return an access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  },
                  "required": [
                    "access_token"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Cannot register as an admin"
          },
          "409": {
            "description": "Email already in use"
          }
        },
        "summary": "Register a new user and return an access token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/google/code": {
      "post": {
        "operationId": "AuthController_loginWithGoogleCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful with Google",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  },
                  "required": [
                    "access_token"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid Google code or code_verifier"
          }
        },
        "summary": "Authenticate with Google authorization code + PKCE",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/github": {
      "post": {
        "operationId": "AuthController_loginWithGithub",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GithubLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful with GitHub",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  },
                  "required": [
                    "access_token"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired GitHub code"
          }
        },
        "summary": "Authenticate with GitHub OAuth authorization code",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/microsoft": {
      "post": {
        "operationId": "AuthController_loginWithMicrosoft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MicrosoftLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful with Microsoft",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  },
                  "required": [
                    "access_token"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid Microsoft ID token"
          }
        },
        "summary": "Authenticate with Microsoft ID token",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Access token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  },
                  "required": [
                    "access_token"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Refresh token missing or invalid"
          }
        },
        "summary": "Refresh the access token using refresh token cookie",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/password": {
      "patch": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully"
          },
          "400": {
            "description": "Current password required for non-OAuth accounts"
          },
          "401": {
            "description": "Current password incorrect"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Change password, OAuth users can set one without providing a current password",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "End the session and clear the refresh token cookie",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Naucto API",
    "description": "The Naucto API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT : Bearer <token>"
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "user@example.com"
          },
          "username": {
            "type": "string",
            "description": "User username",
            "example": "xX_DarkGamer_Xx"
          },
          "nickname": {
            "type": "string",
            "description": "User nick name",
            "example": "JohnDoe"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "username",
          "password"
        ]
      },
      "GoogleCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Google authorization code"
          },
          "codeVerifier": {
            "type": "string",
            "description": "PKCE code verifier"
          }
        },
        "required": [
          "code",
          "codeVerifier"
        ]
      },
      "GithubLoginDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "GitHub OAuth authorization code",
            "example": "abc123def456"
          }
        },
        "required": [
          "code"
        ]
      },
      "MicrosoftLoginDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Microsoft ID token",
            "example": "eyJhbGciOiJSUzI1NiIs..."
          }
        },
        "required": [
          "token"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Current password (not required for OAuth accounts)"
          },
          "newPassword": {
            "type": "string",
            "description": "New password",
            "minLength": 6
          }
        },
        "required": [
          "newPassword"
        ]
      },
...
    }
  }
}